+++ CHANGES.md
@@ -1,3 +1,6 @@
+# UNRELEASED
+- [NEW] Added option for client to authenticate with IAM token server.
+
 # 3.0.2 (2019-01-07)
 - [FIXED] Remove unnecessary `@types/nano` dependancy.